+++ R/tar_meta_delete.R
@@ -79,7 +79,7 @@ tar_meta_delete <- function(
   }
   if (delete %in% c("all", "cloud")) {
     tar_assert_script(script)
-    options <- tar_option_script(script = script)
+    options <- tar_script_options(script = script)
     old_repository_meta <- tar_options$get_repository_meta()
     old_resources <- tar_options$get_resources()
     on.exit({